perf: optimize analytics JS code (#2157)
